Rename internal file and specify non-Kt-suffixed name
